requirements.txt
#Backend (FastAPI & Google)
fastapi
uvicorn[standard]
pydantic
google-cloud-firestore
google-cloud-storage
google-generativeai
google-cloud-logging
fpdf # For PDF generation
python-dotenv

#Frontend (Streamlit)

streamlit
httpx # To call our FastAPI backend